+++ src/core/Entities/RegistrationEvaluation.php
@@ -2,25 +2,37 @@
 
 namespace MapasCulturais\Entities;
 
+use Doctrine\ORM\Exception\NotSupported;
 use MapasCulturais;
 use MapasCulturais\i;
 use MapasCulturais\Traits;
 use Doctrine\ORM\Mapping as ORM;
 use Doctrine\Persistence\Mapping\MappingException;
 use MapasCulturais\App;
+use MapasCulturais\Exceptions\PermissionDenied;
+use MapasCulturais\Exceptions\WorkflowRequest;
 use ReflectionException;
+use RuntimeException;
 
 /**
  * RegistrationMeta
  *
+ * @property-read \MapasCulturais\Definitions\EvaluationMethod $evaluationMethodDefinition
+ * @property-read EvaluationMethodConfiguration $evaluationMethodConfiguration
+ * @property-read \MapasCulturais\EvaluationMethod $evaluationMethod
+ * @property-read string $resultString
+ * @property-read string $statusString
  * @property-read string $result
  * 
  * @property integer $id
- * @property mexed $result
  * @property object $evaluationData
  * @property Registration $registration
  * @property User $user
  * @property integer $status
+ * @property bool $isTiebreaker
+ * @property string $committee
+ * 
+ * 
  *
  * @ORM\Table(name="registration_evaluation")
  * @ORM\Entity
@@ -122,6 +134,15 @@ class RegistrationEvaluation extends \MapasCulturais\Entity {
     protected $__files;
 
     /**
+     * Nome da comissão avaliadora pela qual o avaliador avaliou essa avaliação
+     * 
+     * @var string
+     *
+     * @ORM\Column(name="committee", type="string", nullable=true)
+     */
+    protected $committee = '';
+
+    /**
      * flag que diz que a avaliação está sendo enviada
      * @var boolean
      */
@@ -131,13 +152,13 @@ class RegistrationEvaluation extends \MapasCulturais\Entity {
         if(empty($this->status)){
             $this->status = self::STATUS_DRAFT;
         }
+
+        if(empty($this->committee)){
+            $registration_valuers = $this->registration->valuers;
+            $this->committee = $registration_valuers[$this->user->id] ?? '';
+        }
         
         parent::save($flush);
-        $app = App::i();
-        $opportunity = $this->registration->opportunity;
-        
-        // cache utilizado pelo endpoint findEvaluations
-        $app->mscache->delete("api:opportunity:{$opportunity->id}:evaluations");
     }
 
     function send($flush = false) {
@@ -176,7 +197,7 @@ class RegistrationEvaluation extends \MapasCulturais\Entity {
 
     /**
      * Returns the Evaluation Method Configuration
-     * @return \MapasCulturais\Definitions\EvaluationMethodConfiguration
+     * @return EvaluationMethodConfiguration
      */
     public function getEvaluationMethodConfiguration() {
         return $this->registration->evaluationMethodConfiguration;
@@ -297,6 +318,20 @@ class RegistrationEvaluation extends \MapasCulturais\Entity {
             return \MapasCulturais\i::__('Avaliação de Inscrição');
     }
 
+    /**
+     * Atualiza os resumos do avaliador
+     * 
+     * @return void 
+     */
+    public function updateValuerSummaries() {
+        /** @var EvaluationMethodConfigurationAgentRelation[] */
+        $relations = $this->evaluationMethodConfiguration->getAgentRelationOfUser($this->user);
+
+        foreach($relations as $relation) {
+            $relation->updateSummary();
+        }
+    }
+
     //============================================================= //
     // The following lines ara used by MapasCulturais hook system.
     // Please do not change them.
@@ -314,6 +349,7 @@ class RegistrationEvaluation extends \MapasCulturais\Entity {
         parent::postPersist($args);
         
         $this->registration->consolidateResult(true, $this);
+        $this->updateValuerSummaries();
     }
 
     /** @ORM\PreRemove */
@@ -323,6 +359,7 @@ class RegistrationEvaluation extends \MapasCulturais\Entity {
         parent::postRemove($args);
         
         $this->registration->consolidateResult(true, $this);
+        $this->updateValuerSummaries();
     }
 
     /** @ORM\PreUpdate */
@@ -332,9 +369,6 @@ class RegistrationEvaluation extends \MapasCulturais\Entity {
         parent::postUpdate($args);
         
         $this->registration->consolidateResult(true, $this);
-    }
-
-    public function getResult() {
-        return $this->result;
+        $this->updateValuerSummaries();
     }
 }