+++ src/modules/Opportunities/Module.php
@@ -157,7 +157,9 @@ class Module extends \MapasCulturais\Module{
 
         $app->hook('entity(<<RegistrationEvaluation|Registration>>).send:after', function() use($app, $distribute_execution_time) {
             /** @var Registration $this */
-            $app->enqueueJob(Jobs\RedistributeCommitteeRegistrations::SLUG, ['evaluationMethodConfiguration' => $this->evaluationMethodConfiguration], $distribute_execution_time);
+            if($emc = $this->evaluationMethodConfiguration) {
+                $app->enqueueJob(Jobs\RedistributeCommitteeRegistrations::SLUG, ['evaluationMethodConfiguration' => $emc], $distribute_execution_time);
+            }
         });
 
         // atualiza o cache dos resumos das fase de avaliação
@@ -175,8 +177,10 @@ class Module extends \MapasCulturais\Module{
             }
         });
 
-        $app->hook("entity(Registration).status(<<*>>)", function() use ($app) {
+        $app->hook("entity(Registration).status(<<*>>)", function() use ($app, $distribute_execution_time) {
             $app->log->debug("Registration {$this->id} status changed to {$this->status}");
+
+            $app->enqueueJob(Jobs\RedistributeCommitteeRegistrations::SLUG, ['evaluationMethodConfiguration' => $this->evaluationMethodConfiguration], $distribute_execution_time);
 
             /** @var Registration $this */
             /** @var Opportunity $opportunity */
@@ -224,6 +228,10 @@ class Module extends \MapasCulturais\Module{
                 $validations['registrationFrom']['required'] = i::__('A data inicial das inscrições é obrigatória');
                 $validations['registrationTo']['required'] = i::__('A data final das inscrições é obrigatória');
                 $validations['shortDescription']['required'] = i::__('A descrição curtá é obrigatória');
+            }
+
+            if (empty($this->controller->data['registrationTo'])) {
+                $validations['registrationFrom']['$this->validateRegistrationDates()'] = i::__('A data inicial das inscrições deve ser menor ou igual a data final');
             }
         });
 
@@ -625,7 +633,7 @@ class Module extends \MapasCulturais\Module{
             }
 
             if ($em = $this->getEvaluationMethodConfiguration()) {
-                $em->getUserRelation($user)->updateSummary(flush: true);
+                $em->getUserRelation($user)->updateSummary();
             }
         });
 
@@ -646,7 +654,7 @@ class Module extends \MapasCulturais\Module{
             $this->enqueueUpdateSummary();
         });
 
-        $app->hook("entity(EvaluationMethodConfiguration).renameAgentRelationGroup:before", function($old_name, $new_name, $relations) {
+        $app->hook("entity(EvaluationMethodConfiguration).renameAgentRelationGroup:before", function($old_name, $new_name, $relations) use($app) {
             /** @var \MapasCulturais\Entities\EvaluationMethodConfiguration $this */
 
             if(isset($this->valuersPerRegistration->{$old_name})) {
@@ -664,6 +672,55 @@ class Module extends \MapasCulturais\Module{
 
                 $this->fetchFields = $registration_filter_config;
             }
+        });
+
+        $app->hook("entity(EvaluationMethodConfiguration).renameAgentRelationGroup:after", function($old_name, $new_name, $relations) use($app) {
+            /** @var \MapasCulturais\Entities\EvaluationMethodConfiguration $this */
+
+            /**
+             * Atualiza o nome da comissão dos avaliadores na coluna valuers da tabela registration
+             */
+            $valuer_user_ids = $this->getValuerUserIds(true);
+            $valuer_user_ids = implode(',', $valuer_user_ids);
+            $conn = $app->em->getConnection();
+            if ($valuer_user_ids){
+                $sql = "SELECT valuer_user_id, registration_id, valuer_committee, valuer_user_id 
+                        FROM evaluations 
+                        WHERE valuer_committee = :committe and opportunity_id = :opportunity_id and valuer_user_id in ($valuer_user_ids)";
+
+                $registrations = $conn->fetchAll($sql, [
+                    'committe' => $old_name,
+                    'opportunity_id' => $this->opportunity->id
+                ]);
+
+                foreach($registrations as $registration) {
+                    $registration = (object) $registration;
+                    $update_sql = "
+                        UPDATE registration 
+                        SET valuers['{$registration->valuer_user_id}'] = to_jsonb(:new_name::VARCHAR)
+                        WHERE id = :registration_id";
+
+                    $conn->executeQuery($update_sql, [
+                        'new_name' => $new_name,
+                        'registration_id' => $registration->registration_id
+                    ]);
+                }
+            }
+
+            /**
+             * Atualiza o nome da comissão dos avaliadores na coluna committe da tabela registration_evaluation
+             */
+            $update_evaluations_sql = "
+                UPDATE registration_evaluation 
+                SET committee = :new_name 
+                WHERE   registration_id in (SELECT id FROM registration WHERE opportunity_id = :opportunity_id)
+                    AND committee = :old_name ";
+
+            $conn->executeQuery($update_evaluations_sql, [
+                'new_name' => $new_name,
+                'old_name' => $old_name,
+                'opportunity_id' => $this->opportunity->id
+            ]);
         });
 
         $app->hook("entity(RegistrationEvaluation).send:after", function() use ($app) {