+++ src/modules/OpportunityWorkplan/components/opportunity-enable-workplan/script.js
@@ -13,7 +13,6 @@ app.component('opportunity-enable-workplan', {
     },
     data() {
         return {
-            entity: this.entity,
             timeOut: null
         }
     },